splitting mahout into api and impl
Former-commit-id: 7dc23b01f27eb658363f5c2a644cf34afc09ef29

+++ management/server/products/solr/solr-ui/src/main/java/org/safehaus/kiskis/mgmt/ui/solr/wizard/Wizard.java
@@ -75,12 +75,12 @@ public class Wizard {
                 break;
             }
             case 2: {
-                progressBar.setValue((float) (step - 1) / (NUMBER_OF_STEPS - 1));
+                progressBar.setValue((step - 1) / (NUMBER_OF_STEPS - 1));
                 component = new ConfigurationStep(this);
                 break;
             }
             case 3: {
-                progressBar.setValue((float) (step - 1) / (NUMBER_OF_STEPS - 1));
+                progressBar.setValue((step - 1) / (NUMBER_OF_STEPS - 1));
                 component = new VerificationStep(this);
                 break;
             }
